align used algorithm to detect java bin to use add javaPath for JettyRunDistro mojo to force a java executable to use
